move iso validation code to iso_validation.cpp
